feat(login): Implement PinInputWidget for improved PIN input handling
